move service layer from shop-database module
