+++ shop-admin-app/src/main/java/ru/geekbrains/service/ProductServiceImpl.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.persist.service;
+package ru.geekbrains.service;
 
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Service;
 import ru.geekbrains.persist.model.Product;
 import ru.geekbrains.persist.repository.ProductRepository;
 import ru.geekbrains.persist.specification.ProductSpecification;
-import ru.geekbrains.persist.service.dto.ProductDto;
-import ru.geekbrains.persist.service.dto.ProductMapper;
+import ru.geekbrains.service.dto.ProductDto;
+import ru.geekbrains.service.dto.ProductMapper;
 
 import java.math.BigDecimal;
 import java.util.Locale;
